Migrate MaterialProperties
