+++ crates/bevy_material/src/render_phase/draw.rs
@@ -1,0 +1,6 @@
+use core::{fmt::Debug, hash::Hash};
+
+// TODO: make this generic?
+/// An identifier for a [`Draw`] function stored in [`DrawFunctions`].
+#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord, Hash)]
+pub struct DrawFunctionId(u32);
